Add platform case to detect arm case on Apple systems.
Apple Sillicon's default triple is "arm-apple-darwin20.x", so the vendor
could be used to catch this oddity where the default for arm64 (aka aarch64)
is "arm" vs "arm64".

This is a conservative approach where the vendor string is not added to
the case matching but instead ad-hoc checked for the `arm)` case. This
is to avoid modifying and potentially breaking any of the other cases.