REV - Enable zgc on macOS aarch64.
This Garbage Collector implementation seems to be dependant on identifying
the CPU ID where a thread is currently running. There is not yet a known
way to obtain this value on userland. The current implementation returns
always 0. If the GC relies on this in a way where its usage could be
degraded or broken this might need to be removed.