Set locs_buf type to avoid a compiler error.
This buf of uninitialized relocInfo instances was previously
kept as an array of doubles. As clang understand that
sizeof(type[]) / sizeof(type) is an attempt to calculate the
buffer size, it fails to see a relationship between double and
relocInfo and throws an error. Based on relocInfo.cpp:69, short
seems to be "expected" to be the size of a relocInfo instance.
This changes the buffer data type and extends its possible
entries to match its previous size [read as (8 * 20) / 2 = 80]