REV - Build bsd_aarch64 by merging os_cpu files from bsd_x86 and linux_aarch64.
This is a well-intentioned but not necesarily educated attempt to build
the os-arch specific files for bsd_aarch, by combining the os specific
parts of the bsd_x86 files and the arch specific parts of linux_aarch64.
While the result naively seems to be "functional" there are many low
level specifics that might very well not being catched by tests that
need to be verified.

This needs to be revisited, sanitized, tested and optimized.