FIX - Explicitly use JNF to build Native JDK.
JNF is no longer provided for arm64 builds, its tbd is still defined
for x86_64 tho. To avoid this builds should reference to a static build
of the framework or embed and distribute a dynamic version of it.

This needs to be selectable through a configure setting where a path
to the JNF framework is provided.