FIX - Current CPU ID seems to be unaccessible.
There's no equivalent cpuid instruction, and the value that
resides in MPIDR_EL1 can only be read with a MRS privileged
instruction. This doesn't seem to be accessible at all in
Apple Sillicon systems. As an additionaly note, even Rosetta
that strictly translates operations as close as possible,
is always returning a CPU identifer 0.

FB8687024